pyproject.toml
[tool.poetry]
name = "insta_merge_stories"
version = "0.1.0"
description = "Group images together."
authors = ["Rodrigo Azevedo <rodrig4zevedo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.6.9"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"